+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.20.0] - 2022-01-09
+
+### Added
+
+* Formula for initial TANF eligibility.
+* Two new variables: `tanf_gross_earned_income` and `tanf_gross_unearned_income`.
+* Variable & parameter for `initial_employment_deduction`.
+* Integration tests for TANF cash aid from TANF IL website.
+
+### Changed
+
+* `tanf_countable_income` now includes unearned income and earned income deduction.
+
 ## [0.19.3] - 2022-01-08
 
 ### Changed